feat: Add DataTableDropdown component
Add a new component DataTableDropdown.vue to the MyPublications/Components/dataTable directory. This component includes a dropdown menu triggered by a button with an icon. The dropdown menu contains two options: "Opcion 1" and "Opcion 2". This component is used in the columns.ts file to render a dropdown action button in the table cells.

Refactor the columns.ts file to use the DataTableDropdown component instead of a button element. The DataTableDropdown component is wrapped in a div with the "relative" class to position it correctly. This change improves the user experience by providing a more intuitive and interactive way to perform actions on the table rows.